requirements.txt
tweepy
python-dotenv
requests